requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]==3.12
scikit-learn

-e .
